fix: colocated mutations in single file, updated imports
